Move webapp-common logging utils into logger.py
